% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alaskacounty20.R
\docType{data}
\name{alaskacounty20}
\alias{alaskacounty20}
\title{Alaska}
\format{
A data frame with 30 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(alaskacounty20)
}
\description{
Provides access to the counties of Alaska from the 2020 census redistricting files
}
\keyword{datasets}
