% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arizonacounty20.R
\docType{data}
\name{arizonacounty20}
\alias{arizonacounty20}
\title{Arizona}
\format{
A data frame with 15 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(arizonacounty20)
}
\description{
Provides access to the counties of Arizona from the 2020 census redistricting files
}
\keyword{datasets}
