% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/californiacounty20.R
\docType{data}
\name{californiacounty20}
\alias{californiacounty20}
\title{California}
\format{
A data frame with 58 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(californiacounty20)
}
\description{
Provides access to the counties of California from the 2020 census redistricting files
}
\keyword{datasets}
