% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecticutcounty20.R
\docType{data}
\name{connecticutcounty20}
\alias{connecticutcounty20}
\title{Connecticut}
\format{
A data frame with 8 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(connecticutcounty20)
}
\description{
Provides access to the counties of Connecticut from the 2020 census redistricting files
}
\keyword{datasets}
