% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delawarecounty20.R
\docType{data}
\name{delawarecounty20}
\alias{delawarecounty20}
\title{Delaware}
\format{
A data frame with 3 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(delawarecounty20)
}
\description{
Provides access to the counties of Delaware from the 2020 census redistricting files
}
\keyword{datasets}
