% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districtofcolumbiacounty20.R
\docType{data}
\name{districtofcolumbiacounty20}
\alias{districtofcolumbiacounty20}
\title{DC}
\format{
A data frame with 1 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(districtofcolumbiacounty20)
}
\description{
Provides access to the counties of District of columbia from the 2020 census redistricting files
}
\keyword{datasets}
