% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floridacounty20.R
\docType{data}
\name{floridacounty20}
\alias{floridacounty20}
\title{Florida}
\format{
A data frame with 67 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(floridacounty20)
}
\description{
Provides access to the counties of Florida from the 2020 census redistricting files
}
\keyword{datasets}
