% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawaiicounty20.R
\docType{data}
\name{hawaiicounty20}
\alias{hawaiicounty20}
\title{Hawaii}
\format{
A data frame with 5 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(hawaiicounty20)
}
\description{
Provides access to the counties of Hawaii from the 2020 census redistricting files
}
\keyword{datasets}
