% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idahocounty20.R
\docType{data}
\name{idahocounty20}
\alias{idahocounty20}
\title{Idaho}
\format{
A data frame with 44 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(idahocounty20)
}
\description{
Provides access to the counties of Idaho from the 2020 census redistricting files
}
\keyword{datasets}
