% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illinoiscounty20.R
\docType{data}
\name{illinoiscounty20}
\alias{illinoiscounty20}
\title{Illinois}
\format{
A data frame with 102 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(illinoiscounty20)
}
\description{
Provides access to the counties of Illinois from the 2020 census redistricting files
}
\keyword{datasets}
