% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iowacounty20.R
\docType{data}
\name{iowacounty20}
\alias{iowacounty20}
\title{Iowa}
\format{
A data frame with 99 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(iowacounty20)
}
\description{
Provides access to the counties of Iowa from the 2020 census redistricting files
}
\keyword{datasets}
