% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckycounty20.R
\docType{data}
\name{kentuckycounty20}
\alias{kentuckycounty20}
\title{Kentucky}
\format{
A data frame with 120 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(kentuckycounty20)
}
\description{
Provides access to the counties of Kentucky from the 2020 census redistricting files
}
\keyword{datasets}
