% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louisianacounty20.R
\docType{data}
\name{louisianacounty20}
\alias{louisianacounty20}
\title{Louisiana}
\format{
A data frame with 64 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(louisianacounty20)
}
\description{
Provides access to the counties of Louisiana from the 2020 census redistricting files
}
\keyword{datasets}
