% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marylandcounty20.R
\docType{data}
\name{marylandcounty20}
\alias{marylandcounty20}
\title{Maryland}
\format{
A data frame with 24 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(marylandcounty20)
}
\description{
Provides access to the counties of Maryland from the 2020 census redistricting files
}
\keyword{datasets}
