% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massachussetscounty20.R
\docType{data}
\name{massachussetscounty20}
\alias{massachussetscounty20}
\title{Massachussets}
\format{
A data frame with 14 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(massachussetscounty20)
}
\description{
Provides access to the counties of Massachussets from the 2020 census redistricting files
}
\keyword{datasets}
