% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigancounty20.R
\docType{data}
\name{michigancounty20}
\alias{michigancounty20}
\title{Michigan}
\format{
A data frame with 83 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(michigancounty20)
}
\description{
Provides access to the counties of Michigan from the 2020 census redistricting files
}
\keyword{datasets}
