% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minnesotacounty20.R
\docType{data}
\name{minnesotacounty20}
\alias{minnesotacounty20}
\title{Minnesota}
\format{
A data frame with 87 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(minnesotacounty20)
}
\description{
Provides access to the counties of Minnesota from the 2020 census redistricting files
}
\keyword{datasets}
