% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misssissipicounty20.R
\docType{data}
\name{mississipicounty20}
\alias{mississipicounty20}
\title{Mississipi}
\format{
A data frame with 82 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(mississipicounty20)
}
\description{
Provides access to the counties of Mississipi from the 2020 census redistricting files
}
\keyword{datasets}
