% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missouricounty20.R
\docType{data}
\name{missouricounty20}
\alias{missouricounty20}
\title{Missouri}
\format{
A data frame with 115 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(missouricounty20)
}
\description{
Provides access to the counties of Missouri from the 2020 census redistricting files
}
\keyword{datasets}
