% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montanacounty20.R
\docType{data}
\name{montanacounty20}
\alias{montanacounty20}
\title{Montana}
\format{
A data frame with 56 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(montanacounty20)
}
\description{
Provides access to the counties of Montana from the 2020 census redistricting files
}
\keyword{datasets}
