% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebraskacounty20.R
\docType{data}
\name{nebraskacounty20}
\alias{nebraskacounty20}
\title{Nebraska}
\format{
A data frame with 93 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(nebraskacounty20)
}
\description{
Provides access to the counties of Nebraska from the 2020 census redistricting files
}
\keyword{datasets}
