% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newhampshirecounty20.R
\docType{data}
\name{newhampshirecounty20}
\alias{newhampshirecounty20}
\title{New Hampshire}
\format{
A data frame with 10 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}10
}
\usage{
data(newhampshirecounty20)
}
\description{
Provides access to the counties of New Hampshire from the 2020 census redistricting files
}
\keyword{datasets}
