% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newjerseycounty20.R
\docType{data}
\name{newjerseycounty20}
\alias{newjerseycounty20}
\title{New Jersey}
\format{
A data frame with 21 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newjerseycounty20)
}
\description{
Provides access to the counties of New Jersey from the 2020 census redistricting files
}
\keyword{datasets}
