% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmexicocounty20.R
\docType{data}
\name{newmexicocounty20}
\alias{newmexicocounty20}
\title{New Mexico}
\format{
A data frame with 33 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newmexicocounty20)
}
\description{
Provides access to the counties of New Mexico from the 2020 census redistricting files
}
\keyword{datasets}
