% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northcarolinacounty20.R
\docType{data}
\name{northcarolinacounty20}
\alias{northcarolinacounty20}
\title{North Carolina}
\format{
A data frame with 100 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(northcarolinacounty20)
}
\description{
Provides access to the counties of North Carolina from the 2020 census redistricting files
}
\keyword{datasets}
