% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northdakotacounty20.R
\docType{data}
\name{northdakotacounty20}
\alias{northdakotacounty20}
\title{North Dakota}
\format{
A data frame with 100 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(northdakotacounty20)
}
\description{
Provides access to the counties of North Dakota from the 2020 census redistricting files
}
\keyword{datasets}
