% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohiocounty20.R
\docType{data}
\name{ohiocounty20}
\alias{ohiocounty20}
\title{Ohio}
\format{
A data frame with 88 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(ohiocounty20)
}
\description{
Provides access to the counties of Ohio from the 2020 census redistricting files
}
\keyword{datasets}
