% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oklahomacounty20.R
\docType{data}
\name{oklahomacounty20}
\alias{oklahomacounty20}
\title{Oklahoma}
\format{
A data frame with 77 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(oklahomacounty20)
}
\description{
Provides access to the counties of Oklahma from the 2020 census redistricting files
}
\keyword{datasets}
