% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pennsylvaniacounty20.R
\docType{data}
\name{pennsylvaniacounty20}
\alias{pennsylvaniacounty20}
\title{Pennsylvania}
\format{
A data frame with 67 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(pennsylvaniacounty20)
}
\description{
Provides access to the counties of Pennsylvania from the 2020 census redistricting files
}
\keyword{datasets}
