% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southcarolinacounty20.R
\docType{data}
\name{southcarolinacounty20}
\alias{southcarolinacounty20}
\title{South Carolina}
\format{
A data frame with 46 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(southcarolinacounty20)
}
\description{
Provides access to the counties of South Carolina from the 2020 census redistricting files
}
\keyword{datasets}
