% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southdakotacounty20.R
\docType{data}
\name{southdakotacounty20}
\alias{southdakotacounty20}
\title{South Dakota}
\format{
A data frame with 66 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(southdakotacounty20)
}
\description{
Provides access to the counties of South Dakota from the 2020 census redistricting files
}
\keyword{datasets}
