% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennessecounty20.R
\docType{data}
\name{tennessecounty20}
\alias{tennessecounty20}
\title{Tennesse}
\format{
A data frame with 95 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(tennessecounty20)
}
\description{
Provides access to the counties of Tennesse from the 2020 census redistricting files
}
\keyword{datasets}
