% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texascounty20.R
\docType{data}
\name{texascounty20}
\alias{texascounty20}
\title{Texas}
\format{
A data frame with 254 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(texascounty20)
}
\description{
Provides access to the counties of Texas from the 2020 census redistricting files
}
\keyword{datasets}
