% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscensuscounty20-package.R
\docType{package}
\name{uscensuscounty20-package}
\alias{uscensuscounty20}
\alias{uscensuscounty20-package}
\title{uscensuscounty20: Provides access to the 2020 census county data for various states}
\description{
Provides the access to redistricting files US Census 2020 at the county level for various states
}
\author{
\strong{Maintainer}: shreshtha modi \email{modishreshtha48@gmail.com}

}
\keyword{internal}
