% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virginiacounty20.R
\docType{data}
\name{virginacounty20}
\alias{virginacounty20}
\title{Virginia}
\format{
A data frame with 133 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(virginacounty20)
}
\description{
Provides access to the counties of Virginia from the 2020 census redistricting files
}
\keyword{datasets}
