% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washingtoncounty20.R
\docType{data}
\name{washingtoncounty20}
\alias{washingtoncounty20}
\title{Washington}
\format{
A data frame with 39 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(washingtoncounty20)
}
\description{
Provides access to the counties of Washington from the 2020 census redistricting files
}
\keyword{datasets}
