% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisconsincounty20.R
\docType{data}
\name{wisconsincounty20}
\alias{wisconsincounty20}
\title{Wisconsin}
\format{
A data frame with 23 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(wisconsincounty20)
}
\description{
Provides access to the counties of Wisconsin from the 2020 census redistricting files
}
\keyword{datasets}
