% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyomingcounty20.R
\docType{data}
\name{wyomingcounty20}
\alias{wyomingcounty20}
\title{Wyoming}
\format{
A data frame with 23 rows and 398 variables:
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(wyomingcounty20)
}
\description{
Provides access to the counties of Wyoming from the 2020 census redistricting files
}
\keyword{datasets}
